filebrowser middleware removal gitgnore update vite.config update server code reorganization: api code inside api folder bridge handlers inside bridge folder frontend static inside web/frontend folder
