+++ backend/cmd/server/api/fs.go
@@ -1,4 +1,4 @@
-package system
+package api
 
 import (
 	"net/http"
