+++ backend/cmd/server/api/host.go
@@ -1,4 +1,4 @@
-package system
+package api
 
 import (
 	"net/http"
