+++ backend/cmd/server/api/motherboard.go
@@ -1,4 +1,4 @@
-package system
+package api
 
 import (
 	"fmt"
