+++ backend/cmd/server/api/network.go
@@ -1,4 +1,4 @@
-package system
+package api
 
 import (
 	"fmt"
