+++ backend/cmd/server/api/services.go
@@ -1,4 +1,4 @@
-package system
+package api
 
 import (
 	"bufio"
