+++ backend/cmd/server/api/updates.go
@@ -1,4 +1,4 @@
-package system
+package api
 
 import (
 	"bufio"
