+++ backend/cmd/server/bridge/handlers/updates/routes.go
