+++ backend/cmd/server/bridge/handlers/updates/updates_history.go
