+++ frontend/vite.config.ts
@@ -23,7 +23,7 @@ export default defineConfig(({ command }) => {
       target: "es2017",
       chunkSizeWarningLimit: 2000,
       manifest: true,
-      outDir: "../backend/cmd/server/frontend",
+      outDir: "../backend/cmd/server/webfrontend",
       emptyOutDir: true,
       minify: "esbuild",
     },
